+++ be/src/main/java/kr/codesquad/secondhand/api/address/service/AddressService.java
@@ -2,7 +2,7 @@ package kr.codesquad.secondhand.api.address.service;
 
 import kr.codesquad.secondhand.api.address.domain.Address;
 import kr.codesquad.secondhand.api.address.repository.AddressRepositoryImpl;
-import kr.codesquad.secondhand.api.member.dto.AddressSliceResponse;
+import kr.codesquad.secondhand.api.member.dto.response.AddressSliceResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
