+++ be/src/main/java/kr/codesquad/secondhand/api/member/dto/response/MemberProfileResponse.java
@@ -1,4 +1,4 @@
-package kr.codesquad.secondhand.api.member.dto;
+package kr.codesquad.secondhand.api.member.dto.response;
 
 import kr.codesquad.secondhand.api.member.domain.Member;
 import lombok.Getter;
